Fixed type mismatch (int -> long) when validating a Tweet id in validateId method.

+++ core_java/twitter/src/main/java/ca/jrvs/apps/twitter/service/TwitterService.java
@@ -75,7 +75,7 @@ public class TwitterService implements Service {
    */
   void validateId(String id) {
     try {
-      Integer.parseInt(id);
+      Long.parseLong(id);
     } catch (NumberFormatException e) {
       throw new IllegalArgumentException("Invalid id format", e);
     }
